requirements.txt
python-dotenv
openai
flux_led